+++ iridescent-tetra-expansion-mod/src/main/java/com/iridescentcraft/modspells/item/ModularArsSpellBookItem.java
@@ -144,13 +144,7 @@ public class ModularArsSpellBookItem extends SpellBook implements IModularItem {
             ItemStack stack, int amount, T entity, java.util.function.Consumer<T> onBroken) {
         se.mickelus.tetra.event.ModularItemDamageEvent event =
                 new se.mickelus.tetra.event.ModularItemDamageEvent(entity, stack, amount);
-        try {
-            net.minecraftforge.common.MinecraftForge.EVENT_BUS.post(event);
-        } catch (ClassCastException cce) {
-            org.apache.logging.log4j.LogManager.getLogger("iridescent_modular_spells").warn(
-                    "[icraft] ModularItemDamageEvent listener threw CCE ({}). Skipping that listener's contribution; later listeners on this event were not invoked.",
-                    cce.toString());
-        }
+        com.iridescentcraft.reforging.event.ModularDamageBus.safePost(event, "spellbook_ars", "iridescent_modular_spells");
         int actualAmount = event.getAmount();
         try {
             actualAmount = se.mickelus.tetra.effect.BloodboundEffect.reduceDamage(stack, entity, actualAmount);