+++ iridescent-tetra-expansion-mod/src/main/java/com/iridescentcraft/reforging/item/ItemModularWand.java
@@ -73,13 +73,7 @@ public class ItemModularWand extends Item implements IModularItem {
     public <T extends LivingEntity> int damageItem(ItemStack stack, int amount, T entity, Consumer<T> onBroken) {
         se.mickelus.tetra.event.ModularItemDamageEvent event =
                 new se.mickelus.tetra.event.ModularItemDamageEvent(entity, stack, amount);
-        try {
-            net.minecraftforge.common.MinecraftForge.EVENT_BUS.post(event);
-        } catch (ClassCastException cce) {
-            org.apache.logging.log4j.LogManager.getLogger("iridescent_reforging").warn(
-                    "[icraft] ModularItemDamageEvent listener threw CCE on wand ({}). Continuing.",
-                    cce.toString());
-        }
+        com.iridescentcraft.reforging.event.ModularDamageBus.safePost(event, "wand", "iridescent_reforging");
         int actualAmount = event.getAmount();
         try {
             actualAmount = se.mickelus.tetra.effect.BloodboundEffect.reduceDamage(stack, entity, actualAmount);